Refs #9149 - Remove CustomDialog for CatalogPublish algorithm.
Instead use IPropertySetting objects to handle property complexities.

Eagle-eyed code reviewers will note that these changes don't work very
well with composite catalogs, but neither did the custom algorithm dialog
in its previous form.  This will need to be revisited when composite
catalogs are finished.  In particular, what should be the allowed values
of the InvestigationNumber property given multiple session IDs?  How
should the allowed values (as well as the default value) of the SessionID
property change as well?

Also, for now, "FileName" and "InputWorkspace" are back to being on two
separate rows in the dialog.