Refs #12775. New instrument definition for POLDI 2013
This instrument definition does not need any additional calculations for detector positions, as opposed to the old one. As a trade-off, to stay compatible with previous Mantid versions, the old detector definition is still in the file. Once the detector at POLDI is changed (end of 2016), instrument definition changes anyway and the new files don't need this legacy-solution anymore.